+++ VenmoPayments/src/main/java/com/paypal/android/venmopayments/VenmoClient.kt
@@ -1,0 +1,4 @@
+package com.paypal.android.venmopayments
+
+class VenmoClient {
+}
